requirements.txt
streamlit>=1.30.0
qrcode[pil]>=7.4.0
pyzbar>=0.1.9
pillow>=10.0.0
pytz>=2023.3
pandas>=2.0.0
requests>=2.31.0
